+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
+## [1.57.0] - 2018-06-XX
+-   Add AdminService::getContactInfos method to retrieve informations about a user (need to be loggued in as a admin) 
+-   Add userInfo1 and userInfo2 value in Contact but it can only be filed by roster and AdminService::getContactInfos methods
+
 ## [1.56.0] - 2018-05-28
 -   Add TelephonyService::deflectCall method to deflect a call to an other telephone number  
 -   Update TelephonyService doc.